+++ services/serverbackup/src/main/java/cloud/stackit/sdk/serverbackup/model/BackupProperties.java
@@ -25,11 +25,11 @@ import com.google.gson.stream.JsonWriter;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Set;
 
 /** BackupProperties */
 @javax.annotation.Generated(
@@ -119,6 +119,50 @@ public class BackupProperties {
 		this.volumeIds = volumeIds;
 	}
 
+	/**
+	 * A container for additional, undeclared properties. This is a holder for any undeclared
+	 * properties as specified with the 'additionalProperties' keyword in the OAS document.
+	 */
+	private Map<String, Object> additionalProperties;
+
+	/**
+	 * Set the additional (undeclared) property with the specified name and value. If the property
+	 * does not already exist, create it otherwise replace it.
+	 *
+	 * @param key name of the property
+	 * @param value value of the property
+	 * @return the BackupProperties instance itself
+	 */
+	public BackupProperties putAdditionalProperty(String key, Object value) {
+		if (this.additionalProperties == null) {
+			this.additionalProperties = new HashMap<String, Object>();
+		}
+		this.additionalProperties.put(key, value);
+		return this;
+	}
+
+	/**
+	 * Return the additional (undeclared) property.
+	 *
+	 * @return a map of objects
+	 */
+	public Map<String, Object> getAdditionalProperties() {
+		return additionalProperties;
+	}
+
+	/**
+	 * Return the additional (undeclared) property with the specified name.
+	 *
+	 * @param key name of the property
+	 * @return an object
+	 */
+	public Object getAdditionalProperty(String key) {
+		if (this.additionalProperties == null) {
+			return null;
+		}
+		return this.additionalProperties.get(key);
+	}
+
 	@Override
 	public boolean equals(Object o) {
 		if (this == o) {
@@ -130,12 +174,13 @@ public class BackupProperties {
 		BackupProperties backupProperties = (BackupProperties) o;
 		return Objects.equals(this.name, backupProperties.name)
 				&& Objects.equals(this.retentionPeriod, backupProperties.retentionPeriod)
-				&& Objects.equals(this.volumeIds, backupProperties.volumeIds);
+				&& Objects.equals(this.volumeIds, backupProperties.volumeIds)
+				&& Objects.equals(this.additionalProperties, backupProperties.additionalProperties);
 	}
 
 	@Override
 	public int hashCode() {
-		return Objects.hash(name, retentionPeriod, volumeIds);
+		return Objects.hash(name, retentionPeriod, volumeIds, additionalProperties);
 	}
 
 	@Override
@@ -145,6 +190,9 @@ public class BackupProperties {
 		sb.append("    name: ").append(toIndentedString(name)).append("\n");
 		sb.append("    retentionPeriod: ").append(toIndentedString(retentionPeriod)).append("\n");
 		sb.append("    volumeIds: ").append(toIndentedString(volumeIds)).append("\n");
+		sb.append("    additionalProperties: ")
+				.append(toIndentedString(additionalProperties))
+				.append("\n");
 		sb.append("}");
 		return sb.toString();
 	}
@@ -185,17 +233,6 @@ public class BackupProperties {
 						String.format(
 								"The required field(s) %s in BackupProperties is not found in the empty JSON string",
 								BackupProperties.openapiRequiredFields.toString()));
-			}
-		}
-
-		Set<Map.Entry<String, JsonElement>> entries = jsonElement.getAsJsonObject().entrySet();
-		// check to see if the JSON string contains additional fields
-		for (Map.Entry<String, JsonElement> entry : entries) {
-			if (!BackupProperties.openapiFields.contains(entry.getKey())) {
-				throw new IllegalArgumentException(
-						String.format(
-								"The field `%s` in the JSON string is not defined in the `BackupProperties` properties. JSON: %s",
-								entry.getKey(), jsonElement.toString()));
 			}
 		}
 
@@ -243,6 +280,30 @@ public class BackupProperties {
 						public void write(JsonWriter out, BackupProperties value)
 								throws IOException {
 							JsonObject obj = thisAdapter.toJsonTree(value).getAsJsonObject();
+							obj.remove("additionalProperties");
+							// serialize additional properties
+							if (value.getAdditionalProperties() != null) {
+								for (Map.Entry<String, Object> entry :
+										value.getAdditionalProperties().entrySet()) {
+									if (entry.getValue() instanceof String)
+										obj.addProperty(entry.getKey(), (String) entry.getValue());
+									else if (entry.getValue() instanceof Number)
+										obj.addProperty(entry.getKey(), (Number) entry.getValue());
+									else if (entry.getValue() instanceof Boolean)
+										obj.addProperty(entry.getKey(), (Boolean) entry.getValue());
+									else if (entry.getValue() instanceof Character)
+										obj.addProperty(
+												entry.getKey(), (Character) entry.getValue());
+									else {
+										JsonElement jsonElement = gson.toJsonTree(entry.getValue());
+										if (jsonElement.isJsonArray()) {
+											obj.add(entry.getKey(), jsonElement.getAsJsonArray());
+										} else {
+											obj.add(entry.getKey(), jsonElement.getAsJsonObject());
+										}
+									}
+								}
+							}
 							elementAdapter.write(out, obj);
 						}
 
@@ -250,7 +311,40 @@ public class BackupProperties {
 						public BackupProperties read(JsonReader in) throws IOException {
 							JsonElement jsonElement = elementAdapter.read(in);
 							validateJsonElement(jsonElement);
-							return thisAdapter.fromJsonTree(jsonElement);
+							JsonObject jsonObj = jsonElement.getAsJsonObject();
+							// store additional fields in the deserialized instance
+							BackupProperties instance = thisAdapter.fromJsonTree(jsonObj);
+							for (Map.Entry<String, JsonElement> entry : jsonObj.entrySet()) {
+								if (!openapiFields.contains(entry.getKey())) {
+									if (entry.getValue().isJsonPrimitive()) { // primitive type
+										if (entry.getValue().getAsJsonPrimitive().isString())
+											instance.putAdditionalProperty(
+													entry.getKey(), entry.getValue().getAsString());
+										else if (entry.getValue().getAsJsonPrimitive().isNumber())
+											instance.putAdditionalProperty(
+													entry.getKey(), entry.getValue().getAsNumber());
+										else if (entry.getValue().getAsJsonPrimitive().isBoolean())
+											instance.putAdditionalProperty(
+													entry.getKey(),
+													entry.getValue().getAsBoolean());
+										else
+											throw new IllegalArgumentException(
+													String.format(
+															"The field `%s` has unknown primitive type. Value: %s",
+															entry.getKey(),
+															entry.getValue().toString()));
+									} else if (entry.getValue().isJsonArray()) {
+										instance.putAdditionalProperty(
+												entry.getKey(),
+												gson.fromJson(entry.getValue(), List.class));
+									} else { // JSON object
+										instance.putAdditionalProperty(
+												entry.getKey(),
+												gson.fromJson(entry.getValue(), HashMap.class));
+									}
+								}
+							}
+							return instance;
 						}
 					}.nullSafe();
 		}